rename to comparison_funcs
