+++ ijzer/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod prelude {
-    pub use ijzer_lib::num_funcs::*;
+    pub use ijzer_lib::comparison_funcs::*;
     pub use ijzer_lib::tensor::Tensor;
     pub use ijzer_macro::ijzer;
     pub use num_traits::{pow::Pow, Float};
